requirements.txt
apache-airflow==2.9.3
kafka-python==2.0.2
pandas==2.2.2
requests==2.32.3
SQLAlchemy==2.0.32